\name{NGBVS-package}
\alias{NGBVS-package}
\docType{package}
\title{Bayesian Variable Selection for SNP Data using Normal-Gamma}

\description{
The NGBVS package provides posterior distribution of case-control fine-mapping. Specifically Bayesian variable selection for Single-Nucleotide Polymorphism (SNP) data using the Normal-Gamma prior.
}

\details{
\tabular{ll}{
Package: \tab NG\cr
Type: \tab Package\cr
Version: \tab 0.3.0 \cr
Date: \tab 2022-09-112\cr
License: \tab GPL-2\cr
}
}

\author{
Abdulaziz Alenazi \email{a.alenazi@nbu.edu.sa}
}

\section{Maintainers}{
Abdulaziz Alenazi \email{a.alenazi@nbu.edu.sa}
}

%\references{
%}

%\note{
%Acknowledgments:
%}


